backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
python-multipart
gTTS
pydub
faster-whisper
numpy
